+++ jina/peapods/peas/gateway_pea/__init__.py
@@ -1,0 +1,2 @@
+from .grpc import GatewayPea
+from .rest import RESTGatewayPea
